backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
yfinance==0.2.28
alpha-vantage==2.3.1
stripe==7.8.0
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
prometheus-client==0.19.0
structlog==23.2.0
influxdb-client==1.39.0
kafka-python==2.0.2
alpaca-trade-api==3.1.1
ib-insync==0.9.86
ccxt==4.1.54